+++ srcs/parseur/lum/ft_fill_lum_next.c
@@ -18,11 +18,11 @@ static void		ft_fill_info_lum6(t_obj *obj, t_e *e)
 		&& e->split[2] != NULL && e->split[3] != NULL)
 	{
 		e->vobject.translate_xyz++;
-		e->tmp = atof(e->split[1]);
+		e->tmp = stof(e->split[1]);
 		obj->transform.transl.x = e->tmp;
-		e->tmp = atof(e->split[2]);
+		e->tmp = stof(e->split[2]);
 		obj->transform.transl.y = e->tmp;
-		e->tmp = atof(e->split[3]);
+		e->tmp = stof(e->split[3]);
 		obj->transform.transl.z = e->tmp;
 	}
 	else
@@ -36,15 +36,15 @@ void			ft_fill_info_lum5(t_obj *obj, t_e *e)
 		&& e->split[2] != NULL && e->split[3] != NULL)
 	{
 		e->vobject.rotation_xyz++;
-		e->tmp = atof(e->split[1]);
+		e->tmp = stof(e->split[1]);
 		if (e->tmp < -360 || e->tmp > 360)
 			ft_puterror(e, "Wrong Info in Object - rotation_x");
 		obj->transform.rot.x = e->tmp;
-		e->tmp = atof(e->split[2]);
+		e->tmp = stof(e->split[2]);
 		if (e->tmp < -360 || e->tmp > 360)
 			ft_puterror(e, "Wrong Info in Object - rotation_y");
 		obj->transform.rot.y = e->tmp;
-		e->tmp = atof(e->split[3]);
+		e->tmp = stof(e->split[3]);
 		if (e->tmp < -360 || e->tmp > 360)
 			ft_puterror(e, "Wrong Info in Object - rotation_z");
 		obj->transform.rot.z = e->tmp;
